COMP: fix compile warnings
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=9335
git-svn-id: http://svn.slicer.org/Slicer4/trunk@9335 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Modules/FetchMI/vtkFetchMIParserXND.cxx
@@ -177,7 +177,7 @@ int vtkFetchMIParserXND::ParseMetadataPostResponse( const char *filename )
     }
 
   int found = 0;
-  unsigned int pos;
+  size_t pos;
   while ( !pfile.eof() )
     {
     line.clear();
